feat: pipeline invoking
